+++ src/mem_watch.cc
