+++ src/strategies-provided.cc
